requirements.txt
# Telegram Agent - Requirements
# Python 3.8+ required

# Telegram Bot API
python-telegram-bot==20.8

# Timezone support
pytz==2024.1

# Optional: For enhanced logging
colorlog==6.8.2